API_Produits/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy
pydantic>=1.10.7
pytest>=7.3.1
httpx>=0.24.0  # Pour FastAPI TestProduit
pylint
python-dotenv
requests
pika>=1.3.1
pydantic[email]
pyyaml
psycopg2-binary